+++ AgctorSDK.Host/Program.cs
@@ -89,8 +89,10 @@ if (app.Environment.IsDevelopment())
 {
     app.UseCors("AllowAll");
 }
-
-app.UseHttpsRedirection();
+else
+{
+    app.UseHttpsRedirection();
+}
 app.UseAuthorization();
 app.MapControllers();